% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_filter.R
\name{expression_filter}
\alias{expression_filter}
\title{expression_filter}
\usage{
expression_filter(
  stats,
  clusters,
  mito = 3,
  ribo = 3,
  min.significant = 1,
  min.target.pct = 0.3,
  max.background.pct = 0.7,
  min.diff.pct = 0.2,
  min.de.frac = 0.01,
  min.significance.level = NULL,
  plot = TRUE
)
}
\arguments{
\item{stats}{A data frame containing expression metrics for clusters. See \link{expression_metrics}}

\item{clusters}{A data frame mapping barcodes to clusters. See \link{expression_metrics}}

\item{mito}{NULL or an interger indicating how many deviations a cluster must be to be removed [default = 3].}

\item{ribo}{NULL or an interger indicating how many deviations a cluster must be to be removed [default = 3].}

\item{min.significant}{A number indicating the minimum number of significant genes a cluster must have to pass filtering [default = 1].}

\item{min.target.pct}{A number indicating the minimum mean fraction of barcodes in a cluster that expresses top.n marker genes [default = 0.3].}

\item{max.background.pct}{A number indicating the maximum mean fraction of barcodes outside a cluster that expresses top.n marker genes [default = 0.8].}

\item{min.diff.pct}{A number indicating the minimum difference in percent of barcodes within versus outside a cluster that expresses top.n marker genes [default = 0.2].}

\item{min.de.frac}{A number indicating the minimum fraction of tested genes that must be significant for a cluster to pass filtering [default = 0.01]}

\item{min.significance.level}{A number indicating the minimum significance level of the most signficiant marker gene for a cluster to pass filtering. Set to NULL to identify the threshold automatically [default = NULL].}

\item{plot}{A boolean (TRUE or FALSE), which indicates if a plot should be returned [default = TRUE]}
}
\value{
A vector of valid barcodes
}
\description{
Function to filter on expression metrics
}
