% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_metrics.R
\name{expression_metrics}
\alias{expression_metrics}
\title{expression_metrics}
\usage{
expression_metrics(
  counts,
  mito,
  ribo,
  nfeats = 5000,
  npcs = 10,
  k.min = 5,
  res.shallow = 0.1,
  top.n = 10
)
}
\arguments{
\item{counts}{A matrix containing counts for barcodes passing quality filtering. See \link{quality_filter}}

\item{mito}{A vector containing names of mitochondrial genes. See \link{quality_metrics}}

\item{ribo}{A vector containing names of ribosomal genes. See \link{quality_metrics}}

\item{nfeats}{A number indicating the number of variable features to use for clustering [default = 5000].}

\item{npcs}{A number indicating the number of singular values to use for clustering [default = 10].}

\item{k.min}{A number indicating the lowest number barcodes in a cluster [default = 5].}

\item{res.shallow}{A number indicating the resolution to use for shallow clustering [default = 0.1].}

\item{top.n}{A number indicating the top genes to use for analysis [default = 10].}
}
\value{
A data frame containing expression metrics
}
\description{
Function to calculate expression metrics
}
