% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_data.R
\docType{data}
\name{human}
\alias{human}
\title{A dataset containing human genes annotations}
\format{
A data frame with 106851 rows and 7 variables:
\describe{
\item{NCBI}{NCBI gene (formerly Entrezgene) ID}
\item{HGNC}{HGNC ID}
\item{Ensembl}{Stable Ensemble gene ID}
\item{Chr}{Chromosome scaffold name}
\item{Symbol}{Official gene name}
\item{Type}{Gene type}
\item{Alias}{Gene synonym/aliases}
}
}
\source{
\url{http://www.ensembl.org/}
}
\usage{
data(human)
}
\description{
A dataset containing human genes annotations
}
\keyword{datasets}
