% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_apoptotic.R
\name{label_apoptotic}
\alias{label_apoptotic}
\title{label_apoptotic}
\usage{
label_apoptotic(
  counts,
  metrics,
  zscore = 3,
  nfeats = 5000,
  npcs = 20,
  anomaly.quantile = 0.15,
  plot = TRUE,
  seed = 42,
  verbose = FALSE
)
}
\arguments{
\item{counts}{A matrix containing counts for barcodes passing the barcode-rank threshold. See \link{rank_barcodes}}

\item{metrics}{A data frame containing quality metrics. See \link{quality_metrics} and \link{quality_filter}}

\item{zscore}{A number indicating the Z score to use for soft labeling of apoptotic cells [default = 3].}

\item{nfeats}{A number indicating the number of variable features to use for clustering [default = 5000].}

\item{npcs}{A number indicating the number of singular values to use for clustering [default = 100].}

\item{anomaly.quantile}{A number indicating the quantile to use for detecting anomalous barcodes [default = 0.1].}

\item{plot}{A boolean (TRUE or FALSE) indicating whether or not to produce plots [default = TRUE].}

\item{seed}{Set a seed for reproducible results. Set to NULL to run in non-deterministic mode [default = 42].}

\item{verbose}{Set to TRUE for verbose mode. Set to FALSE for silent mode [default = TRUE].}
}
\value{
A data frame object that contains the quality metrics of the data and results indicating the apoptotic cells.
}
\description{
Function that uses Isolation Forest to label apoptotic cells.
}
