% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_dead.R
\name{label_dead}
\alias{label_dead}
\title{label_dead}
\usage{
label_dead(
  counts,
  metrics,
  qc.labels,
  cor.threshold = NULL,
  train = TRUE,
  rep = 10,
  n.min = 8,
  n.relabel = 1,
  feature.try = 3,
  verbose = TRUE,
  label.thrs = NULL,
  label.frac = 0.1,
  nfeats = 2000,
  alpha = 0,
  npcs = 100,
  weight = TRUE,
  epochs = 20,
  nfolds = 5,
  nrep = 10,
  fail.weight = 0.2,
  cor.min = 1e-04,
  cor.max = 0.005,
  cor.steps = 50,
  nrep.cor = 10,
  min.dead = 100,
  max.live = 500,
  plot = TRUE,
  bpparam = SerialParam()
)
}
\arguments{
\item{counts}{A matrix containing counts for barcodes passing the barcode-rank threshold. See \link{rank_barcodes}}

\item{metrics}{A data frame containing quality metrics. See \link{quality_metrics} and \link{quality_filter}}

\item{qc.labels}{A named vector of barcodes quality filter labels ("pass" or "fail").}

\item{cor.threshold}{Threshold on correlation to use for feature selection. Set to NULL to automatically identify the threshold [default = NULL].}

\item{train}{A boolean (TRUE or FALSE) indicating whether or not to optimize labels [default = TRUE].}

\item{rep}{An integer indicating the number of times to optimize labels [default = 10].}

\item{n.min}{An integer indicating the number of times a predicted label must be consistent to be retained [default = 8].}

\item{n.relabel}{An integer indicating the number of dead cells to label if soft-labelling identifies no dead cells [default = 1].}

\item{feature.try}{An integer indicating the number of times to reduce the correlation threshold boundaries if no good solutions are found [default = 3].}

\item{verbose}{A boolean (TRUE or FALSE) indicating whether or not to be verbose [default = TRUE].}

\item{label.thrs}{Threshold to use for soft-labeling barcodes as either dead or live. If set to NULL, automatic threshold identification. [default = NULL].}

\item{label.frac}{Maximum fraction of barcodes to soft-label as dead [default = 0.1].}

\item{nfeats}{A number indicating the number of variable features to use for SVD [default = 2000].}

\item{alpha}{Alpha to use in glmnet [default = 0].}

\item{npcs}{A number indicating the number of singular values to calculate [default = 100].}

\item{weight}{A boolean (TRUE or FALSE) indicating whether or not to be weight barcodes during training [default = TRUE].}

\item{epochs}{The maximum number of epochs to train for [default = 20].}

\item{nfolds}{The number of folds to use in cross-validation [default = 5].}

\item{nrep}{The number of replicates to use during train time [default = 10].}

\item{fail.weight}{The weight to assign to barcodes that failed quality control [default = 0.2].}

\item{cor.min}{The minimum threshold to test for correlation-based feature selection [default = 0.0001].}

\item{cor.max}{The maximum threshold to test for correlation-based feature selection [default = 0.005].}

\item{cor.steps}{The number of steps between cor.min and cor.max to test for correlation-based feature selection [default = 50].}

\item{nrep.cor}{The number of replicates to perform when determining the correlation threshold [default = 10].}

\item{min.dead}{The minimum number of dead barcodes to use for training [default = 100].}

\item{max.live}{The maximum number of live barcodes to use for training [default = 500].}

\item{plot}{A boolean (TRUE or FALSE) indicating whether or not to plot the scoring threshold [default = TRUE].}

\item{bpparam}{The backend to use for processing. Set to MulticoreParam() or SnowParam() for parallel processing [default = SeriamParam()].}
}
\value{
A data frame object that contains the quality metrics of the data and results indicating the apoptotic cells.
}
\description{
Identification of dead barcode using a simple heurestic to soft-label followed by label optimization using glmnet and adaptive resampling.
}
