% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_data.R
\docType{data}
\name{mouse}
\alias{mouse}
\title{A dataset containing mouse genes annotations}
\format{
A data frame with 82003 rows and 7 variables:
\describe{
\item{NCBI}{NCBI gene (formerly Entrezgene) ID}
\item{MGI}{MGI ID}
\item{Ensembl}{Stable Ensemble gene ID}
\item{Chr}{Chromosome scaffold name}
\item{Symbol}{Official gene name}
\item{Type}{Gene type}
\item{Alias}{Gene synonym/aliases}
}
}
\source{
\url{http://www.ensembl.org/}
}
\usage{
data(mouse)
}
\description{
A dataset containing mouse genes annotations
}
\keyword{datasets}
