% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_filter.R
\name{quality_filter}
\alias{quality_filter}
\title{quality_filter}
\usage{
quality_filter(
  metrics,
  mito = TRUE,
  distance = TRUE,
  coding = TRUE,
  contrast = FALSE,
  mito.nreps = 10,
  dist.degree = 3,
  dist.thres = "auto",
  coding.threshold = 5,
  plot = TRUE
)
}
\arguments{
\item{metrics}{A data frame containing quality metrics. See \link{quality_metrics}}

\item{mito}{A boolean (TRUE or FALSE) indicating whether or not to filter using a threshold on the fraction of UMIs derived from mitochondrial genes [default = TRUE].}

\item{distance}{A boolean (TRUE or FALSE) indicating whether or not to filter using a threshold on the residuals from robust regression between number of features and number of UMIs [default = TRUE].}

\item{coding}{A boolean (TRUE or FALSE) indicating whether or not to filter using a threshold on the fraction of UMIs derived from protein-coding genes [default = TRUE].}

\item{contrast}{A boolean (TRUE or FALSE) indicating whether or not to filter using a threshold on the contrast fraction [default = FALSE]. See \link{quality_metrics}}

\item{mito.nreps}{A numeric indicating the number of times to repeat threshold identification for mitochondrial filtering [default = 10].}

\item{dist.degree}{A numeric indicating the degree for the polynomial for robust regression [default = 3].}

\item{dist.thres}{The maximum number of standard deviations below the mean that passes the QC. Set to ""auto" to automatically determine the threshold [default = "auto"]. See Details for more information}

\item{coding.threshold}{The maximum number of standard deviations around the mean that passes the QC [default = 5].}

\item{plot}{A boolean (TRUE or FALSE) indicating whether or not to produce plots [default = TRUE].}
}
\value{
A list of vectors containing the mitochondrial threshold, number of barcodes filtered at each step and the final barcodes that pass QC filtering.
}
\description{
Function to filter barcodes based on quality metrics
}
\details{
\strong{Thresholding regression residuals on distance between number of features and number of UMIs}\cr
In our experience, the residuals from polynomial regression analysis of the number of features and number of UMIs almost always produces a left-tailed normal distribution. The default behaviour is to find the smallest number of standard deviations above the mean that includes all data points and use to threshold below the mean.
}
