% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_barcode.R
\name{rank_barcode}
\alias{rank_barcode}
\title{rank_barcode}
\usage{
rank_barcode(
  counts,
  input_type = "UMI",
  threshold = TRUE,
  plot = TRUE,
  plot_format = "none",
  directory = FALSE
)
}
\arguments{
\item{counts}{a matrix containing counts of the genes or UMI}

\item{input_type}{an object to be ranked (e.g., UMI or genes)}

\item{threshold}{a boolean for threshold to be included in the output}

\item{plot}{a boolean}

\item{plot_format}{a format for the output plot (e.g., available choises: pdf, tiff, jpg, png or none)}

\item{directory}{a boolean}
}
\value{
A list or a data frame object that contains ranked barcodes and if chosen a threshold.
}
\description{
Function ranks UMI/genes according to their expression levels
}
