% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_idf.R
\name{tf_idf}
\alias{tf_idf}
\title{tf_idf}
\usage{
tf_idf(
  bcranks,
  lower_threshold = 0,
  max.frac = 1,
  diff.frac = 0.5,
  nsteps.frac = 20,
  plot = TRUE,
  plot_format = "none",
  directory = FALSE,
  ncpus = "detect"
)
}
\arguments{
\item{bcranks}{a dataframe of ranked genes or UMI}

\item{lower_threshold}{a threshold to define barcodes that are within acceptable range}

\item{max.frac}{a maximum fraction}

\item{diff.frac}{the difference for the fractions}

\item{nsteps.frac}{steps taken between the fraction range}

\item{plot}{a boolean}

\item{plot_format}{a format for the output plot (e.g., available choises: pdf, tiff, jpg, png or none)}

\item{directory}{a boolean}

\item{ncpus}{a number of CPUs to be used (default: will be detected identifying user's computer specs)}
}
\value{
A dataframe with statistical metrics from TF-IDF
}
\description{
Function uses 'Term Frequancy - Inverse Document Frequency' statistical measure to select informative genes
}
