% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_data.R
\docType{data}
\name{zebrafish}
\alias{zebrafish}
\title{A dataset containing zebrafish genes annotations}
\format{
A data frame with 67809 rows and 6 variables:
\describe{
\item{NCBI}{NCBI gene (formerly Entrezgene) ID}
\item{Ensembl}{Stable Ensemble gene ID}
\item{Chr}{Chromosome scaffold name}
\item{Symbol}{Official gene name}
\item{Type}{Gene type}
\item{Alias}{Gene synonym/aliases}
}
}
\source{
\url{http://www.ensembl.org/}
}
\usage{
data(zebrafish)
}
\description{
A dataset containing zebrafish genes annotations
}
\keyword{datasets}
