% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_apoptotic.R
\name{label_apoptotic}
\alias{label_apoptotic}
\title{label_apoptotic}
\usage{
label_apoptotic(
  counts,
  metrics,
  qc.labels,
  cor.threshold = NULL,
  verbose = TRUE,
  label.thrs = 13.5,
  nfeats = 2000,
  alpha = 0,
  npcs = 100,
  weight = TRUE,
  epochs = 10,
  nfolds = 5,
  nrep = 50,
  fail.weight = 0.2,
  cor.min = 1e-04,
  cor.max = 0.005,
  cor.steps = 50,
  nrep.cor = 20,
  min.apoptotic = 100,
  max.healthy = 500
)
}
\arguments{
\item{counts}{A matrix containing counts for barcodes passing the barcode-rank threshold. See \link{rank_barcodes}}

\item{metrics}{A data frame containing quality metrics. See \link{quality_metrics} and \link{quality_filter}}

\item{verbose}{Set to TRUE for verbose mode. Set to FALSE for silent mode [default = TRUE].}

\item{nfeats}{A number indicating the number of variable features to use for clustering [default = 5000].}

\item{npcs}{A number indicating the number of singular values to use for clustering [default = 100].}

\item{mitochondrial.threshold}{A mitochondrial threshold identified using quality_filter method [default = 0.5].}

\item{threshold}{A number indicating the score to use for soft labeling of apoptotic cells [default = 0.2].}

\item{seed}{Set a seed for reproducible results. Set to NULL to run in non-deterministic mode [default = 42].}
}
\value{
A data frame object that contains the quality metrics of the data and results indicating the apoptotic cells.
}
\description{
Function that uses Isolation Forest to label apoptotic cells.
}
