% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valiDrops.R
\name{valiDrops}
\alias{valiDrops}
\title{valiDrops}
\usage{
valiDrops(
  counts,
  rank_barcodes = TRUE,
  mitochondrial_clusters = 3,
  ribosomal_clusters = 3,
  label_dead = FALSE,
  status = TRUE,
  stageThree = TRUE,
  ...
)
}
\arguments{
\item{counts}{A matrix containing counts for barcodes passing the barcode-rank threshold.}

\item{rank_barcodes}{A boolean (TRUE or FALSE) indicating whether or not to rank barcodes [default = TRUE].}

\item{mitochondrial_clusters}{NULL or an interger indicating how many deviations a cluster must be to be removed [default = 3].}

\item{ribosomal_clusters}{NULL or an interger indicating how many deviations a cluster must be to be removed [default = 3].}

\item{label_dead}{A boolean (TRUE or FALSE) indicating whether or not to label putative dead cells [default = FALSE].}

\item{status}{A boolean (TRUE or FALSE) indicating whether or not to print the progress of valiDrops [default = TRUE].}

\item{stageThree}{A boolean (TRUE or FALSE) indicating whether or not to run stage 3 of valiDrops [default = TRUE].}

\item{...}{Pass parameters to functions within valiDrops. See \link{rank_barcodes} \link{quality_metrics} \link{quality_filter} \link{expression_metrics} \link{expression_filter} \link{label_dead}}
}
\value{
A data frame containing quality metrics, as well as quality control labels (and if requested, dead labels) for all barcodes passing the rank threshold.
}
\description{
Wrapper function that runs the entire valiDrops flow with default parameters.
}
